feat: update JWT config to support float expiration time
- Change ExpiryHours type in JWTConfig from integer to float for more precise expiration setting
- Update related services to use new float-based expiration calculation
- Add getenvFloatDefault function to read float values from environment variables
- Remove deprecated jwt_test.go file

+++ internal/service/admin/auth/login.go
@@ -59,7 +59,7 @@ func (s *Login) Login(ctx context.Context, req adminAuthModel.LoginRequest, logi
 	response := &adminAuthModel.LoginResponse{
 		AccessToken:  accessToken,
 		RefreshToken: refreshToken,
-		ExpiresIn:    s.jwtConfig.ExpiryHours * 3600,
+		ExpiresIn:    int(s.jwtConfig.ExpiryHours * 3600),
 	}
 
 	return response, nil
